Differential Revision: [D35124883](https://our.internmc.facebook.com/intern/diff/D35124883)

This PR kicks off a refactor of Vulkan API objects to improve configurability and readability, and re-organize owners of Vulkan objects to enable a delegate design pattern.

For context on these changes, the current layout of the Vulkan backend architecture is as follows:
* There is a global instance of the `Runtime` class that holds the `VkInstance` object. This class is responsible for creating the `VkInstance` and enumerating information about the physical devices visible to the `VkInstance`, which are stored in `Adapter` class. Its primary use is to enumerate the physical devices available to the `VkInstance` and select a suitable device to use for computation. 
* There is a global instance of the `Context` class that holds the `VkDevice` object, and contains all structures necessary to execute compute shaders on a GPU (command pool, shader factory, resource allocator, etc.). It is the primary interface for building and submitting GPU command buffers. It is responsible for creating the `VkDevice` from the `VkPhysicalDevice` selected by the global `Runtime` instance, as well as creating all member structures using the `VkDevice`. All calls to the Vulkan API are routed through this `Context` class.
    * The member structures are as follows:
        * `Shader` - creates `VkShader` objects
        * `Pipeline` - creates `VkPipeline` objects
        * `Resource` - allocates `VkBuffer` and `VkImage` objects. This structure is `thread_local`
        * `Descriptor` - allocates `VkDescriptor` objects. This structure is `thread_local`
        * `Command` - allocates `VkCommandBuffer` objects. This structure is `thread_local`
    * The above objects are all associated with a specific `VkDevice`. 
    * Why are some structures `thread_local` and others aren’t? It is to enable multiple Vulkan models to execute simultaneously in separate threads. At the end of each inference, Vulkan resources must be purged to free up GPU memory. Shader and Pipeline objects are tied to specific shader programs so they can be reused between inferences. Buffers/Tetures, Descriptors, and Command Buffers are unique to each dispatch and therefore those pools must be purged at the end of each inference. Therefore, these structures should be associated with a particular execution thread, whereas Shader and Pipeline can just have one global instance tied to the global Context

In order to improve the readability and code organization of the code base. The following structural changes are planned:

* Ownership of `VkDevice` is moved to the `Adapter` class. It will represent the capabilities of one device.
  * Accordingly, `Adapter` will take ownership of the Shader and Pipeline structs, and also be able to create Descriptor, Memory, and Command pools for a particular `Context` class
* The `Runtime` will own `Adapters` for all physical devices visible to the `VkInstance` it owns
* The `Context` class will represent one thread of execution. Instead of owning the `VkDevice`, it will hold an index pointing to an `Adapter` owned by the global `Runtime` instance. Additionally, instead of owning all facilities needed for GPU compute, it will only hold Memory, Descriptor, and Command pools to support one thread of execution, which it will request from the `Adapter` referenced. The `Context` class will itself be made `thread_local`, rather than having a confusing mix of some structures owned by `Context` be stored in `static thread_local` variables.

The above changes will enable a Delegate pattern, allowing each model to hold their own Memory, Descriptor, and Command pools sized exactly to the requirements of the model.

[ghstack-poisoned]

+++ aten/src/ATen/native/vulkan/api/Runtime.h
@@ -44,21 +44,21 @@ class Runtime final {
 
   VkInstance instance() const;
 
-  using Selector = std::function<size_t (const std::vector<Adapter>&)>;
-  size_t init_adapter(const Selector& selector);
+  using Selector = std::function<int32_t (const std::vector<Adapter>&)>;
+  int32_t init_adapter(const Selector& selector);
 
   Adapter* get_adapter_p();
   Adapter& get_adapter();
 
-  Adapter* get_adapter_p(size_t i);
-  Adapter& get_adapter(size_t i);
+  Adapter* get_adapter_p(int32_t i);
+  Adapter& get_adapter(int32_t i);
 
-  size_t default_adapter_i() const;
+  int32_t default_adapter_i() const;
 
  private:
   VkInstance instance_;
   std::vector<Adapter> adapters_;
-  size_t default_adapter_i_;
+  int32_t default_adapter_i_;
 
   VkDebugReportCallbackEXT debug_report_callback_;
 };
@@ -89,15 +89,15 @@ inline Adapter& Runtime::get_adapter() {
   return adapters_[default_adapter_i_];
 }
 
-inline Adapter* Runtime::get_adapter_p(size_t i) {
+inline Adapter* Runtime::get_adapter_p(int32_t i) {
   return &adapters_[i];
 }
 
-inline Adapter& Runtime::get_adapter(size_t i) {
+inline Adapter& Runtime::get_adapter(int32_t i) {
   return adapters_[i];
 }
 
-inline size_t Runtime::default_adapter_i() const {
+inline int32_t Runtime::default_adapter_i() const {
   return default_adapter_i_;
 }
